Created csv_utils.py and migrated the functions
Added
- csv_utils.py and the functions from helper.py

Fixed
- Import errors
